Updates:  - Opção de definir rota entre sua localidade e a localidade do lava jato adicionado.  - Adicionada pergunta se o usuario aceita liberar o acesso a localidade dele, caso esta esteja desativada no android.

+++ app/src/main/java/br/com/nwk/materialdesign/model/LavaJato.java
@@ -1,5 +1,7 @@
 package br.com.nwk.materialdesign.model;
 
+import android.util.Log;
+
 import org.json.JSONException;
 import org.json.JSONObject;
 
@@ -88,8 +90,9 @@ public class LavaJato implements Serializable {
         if(object.has(objectName)){
             try{
                 if(!object.isNull(objectName)){
-                    result = object.getInt(objectName);
-                    //System.out.println(object.getString(objectName));
+                    result = object.getDouble(objectName);
+
+                    //Log.e("Test", )
                 }
 
             }catch (JSONException e){
